+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   schema version is non-zero.
   Fixes issue [#2724](https://github.com/realm/realm-core/issues/2724).
   PR [#2726](https://github.com/realm/realm-core/pull/2726).
+* Fix incorrect results from TableView::find_first().
 
 ### Breaking changes